fix(sync): tuple claim-candidate fetch, CI migration replay, saveAscent normalizeTimestamp
Round-2 review on PR #3555:

1. Cross-source claim candidate fetch now uses a row-value
   (climb_uuid, angle) IN ((...), ...) tuple filter instead of two
   independent IN-lists, which fetched the cartesian product (every listed
   climb at every listed angle). Bounded at ≤100 tuples per statement
   (WRITE_CHUNK_SIZE); the JS pair match stays as belt-and-braces. Test
   renders the WHERE via PgDialect and asserts the tuple shape + that only
   the real pairs are bound.

2. The dedup migration replay now runs in CI: extracted the synthetic
   schema/seed/assertions into @boardsesh/db/testing/dedup-replay and added a
   backend-project vitest harness (dedup-migration-replay.test.ts) that runs
   on every CI backend job against the auto-started test postgres, creating a
   per-worker throwaway database (bs_dedup_replay_w<pool>). The packages/db
   node:test harness keeps the MIGRATION_REPLAY_DB_URL opt-in mode for local
   scratch runs; both share the fixture so they can't drift.

3. 0164/0165 recompute SQL: added the frozen-in-time header — the duplicated
   recompute logic is a deliberate snapshot of recompute.ts at the moment the
   migration shipped (same precedent as merged 0157); sharing code would let
   an already-applied migration's meaning mutate.

4. saveAscent routes climbed_at through normalizeTimestamp (new narrow
   @boardsesh/aurora-sync/normalize-timestamp subpath export) and the comment
   no longer overstates timezone safety: ISO-with-zone passes through
   untouched, an Aurora-style naive string is pinned to UTC instead of
   parsed server-local, and the zoneless-ISO caveat is stated explicitly.
